+++ CHANGES.rst
@@ -13,6 +13,13 @@ Changelog of TRS
 
 - We require python 3.4 now as we use the new "statistics" module.
 
+- Fixed bug that members of a hidden project (like sickness projects) *could*
+  download the full ``.csv`` file with all other peoples' info...
+
+- Fixed bug that the project-is-ended state would be cached. This state
+  changes without the project itself changing, so the cached value would stay
+  the same. Now we just calculate it every time.
+
 
 1.13 (2015-06-18)
 -----------------