+++ CHANGES.rst
@@ -25,6 +25,11 @@ Changelog of TRS
 
 - Added two columns to project csv export.
 
+- Added overview page which makes "booking into the red" more visible, both
+  for hours and for money.
+
+- Made other persons' bookings editable for those that can edit everything.
+
 
 1.13 (2015-06-18)
 -----------------